+++ src/Ladder.java
@@ -2,7 +2,7 @@ import java.awt.Color;
 
 public class Ladder extends GameObject{
 
-	public Ladder(int x, int y, int h, int w) {
+	public Ladder(float x, float y, int h, int w) {
 		super(x, y, h, w);
 		color = Color.cyan;
 		isSolid = false;
